+++ dalesbred/src/main/java/org/dalesbred/transaction/AbstractTransactionManager.java
@@ -34,27 +34,24 @@ public abstract class AbstractTransactionManager implements TransactionManager {
 
     protected abstract <T> T withNewTransaction(@NotNull TransactionCallback<T> callback,
                                                 @NotNull Dialect dialect,
-                                                @NotNull Isolation isolation,
-                                                int retries);
+                                                @NotNull Isolation isolation);
 
     protected abstract <T> T withSuspendedTransaction(@NotNull TransactionCallback<T> callback,
                                                       @NotNull Isolation isolation,
-                                                      @NotNull Dialect dialect,
-                                                      int retries);
+                                                      @NotNull Dialect dialect);
 
     @Override
     public <T> T withTransaction(@NotNull TransactionSettings settings, @NotNull TransactionCallback<T> callback, @NotNull Dialect dialect) {
         Propagation propagation = settings.getPropagation();
         Isolation isolation = settings.getIsolation();
-        int retries = settings.getRetries();
 
         DefaultTransaction existingTransaction = getActiveTransaction().orElse(null);
 
         if (existingTransaction != null) {
             if (propagation == Propagation.REQUIRES_NEW)
-                return withSuspendedTransaction(callback, isolation, dialect, retries);
+                return withSuspendedTransaction(callback, isolation, dialect);
             else if (propagation == Propagation.NESTED)
-                return existingTransaction.nested(retries, callback, dialect);
+                return existingTransaction.nested(callback, dialect);
             else
                 return existingTransaction.join(callback, dialect);
 
@@ -62,7 +59,7 @@ public abstract class AbstractTransactionManager implements TransactionManager {
             if (propagation == Propagation.MANDATORY)
                 throw new NoActiveTransactionException("Transaction propagation was MANDATORY, but there was no existing transaction.");
 
-            return withNewTransaction(callback, dialect, isolation, retries);
+            return withNewTransaction(callback, dialect, isolation);
         }
     }