+++ dalesbred/src/test/java/org/dalesbred/TransactionSettingsTest.java
@@ -37,13 +37,7 @@ public class TransactionSettingsTest {
 
         settings.setPropagation(Propagation.REQUIRED);
         settings.setIsolation(Isolation.REPEATABLE_READ);
-        settings.setRetries(3);
 
-        assertEquals("[propagation=REQUIRED, isolation=REPEATABLE_READ, retries=3]", settings.toString());
-    }
-
-    @Test(expected=IllegalArgumentException.class)
-    public void negativeRetries() {
-        new TransactionSettings().setRetries(-1);
+        assertEquals("[propagation=REQUIRED, isolation=REPEATABLE_READ]", settings.toString());
     }
 }